Refactor : 마켓 서비스 구조 분리
