+++ src/main/java/com/umc/finly/domain/market/service/MarketIndexService.java
@@ -1,37 +1,8 @@
 package com.umc.finly.domain.market.service;
 
-import com.umc.finly.domain.market.converter.MarketIndexConverter;
-import com.umc.finly.domain.market.dto.response.FearGreedResDTO;
 import com.umc.finly.domain.market.dto.response.MarketIndexResDTO;
-import com.umc.finly.domain.market.dto.response.MarketIndicesDTO;
-import com.umc.finly.domain.market.infra.FearGreedIndexProvider;
-import com.umc.finly.domain.market.infra.MarketIndexProvider;
-import lombok.RequiredArgsConstructor;
-import org.springframework.cache.annotation.Cacheable;
-import org.springframework.stereotype.Service;
 
-import java.time.LocalDateTime;
-import java.time.ZoneId;
+public interface MarketIndexService {
 
-@Service
-@RequiredArgsConstructor
-public class MarketIndexService {
-    private final MarketIndexProvider marketIndexProvider;
-    private final FearGreedIndexProvider fearGreedIndexProvider;
-
-    @Cacheable(
-            value = "marketIndex",
-            unless = "#result == null"
-    )
-    public MarketIndexResDTO getMarketIndex() {
-        MarketIndicesDTO indices = marketIndexProvider.getMarketIndices();
-        FearGreedResDTO fearGreed = fearGreedIndexProvider.getFearGreedIndex();
-        LocalDateTime snapshotAt = LocalDateTime.now(ZoneId.of("Asia/Seoul"));
-
-        return MarketIndexConverter.toMarketIndexSnapshot(
-                indices,
-                fearGreed,
-                snapshotAt
-        );
-    }
+    MarketIndexResDTO getMarketIndex();
 }
